Create README.md
The Almond 3 is a “smart home Wi-Fi system” made by Securifi. While IoT products like this are growing in popularity and complexity, they are also putting homes and businesses at risk due to the inherent lack of built-in security. As a router capable of port forwarding and controlling multiple low-energy sensor networks, the Almond 3 gives any successful hacker a dangerous amount of power. With that in mind, I decided to assess the security of each of the Almond’s components: web application, mobile application, networks and hardware. Throughout this paper I detail the methodologies used, vulnerabilities I found, and several ideas for further research.